Proper type conversion

+++ src/1darray/pack.h
@@ -11,7 +11,8 @@ static inline unsigned pack_1darray_size_type(unsigned char buf[],
                                               enum lip_1darray_type type)
 {
     if (size > UINT_MAX / _1darray_item_size(type)) return 0;
-    return lip_pack_ext_size_type(buf, _1darray_item_size(type) * size, type);
+    return lip_pack_ext_size_type(buf, _1darray_item_size(type) * size,
+                                  (uint8_t)type);
 }
 
 #endif
